fin demande et la plus part des create

+++ app/Http/Controllers/CategorieController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers;
 use App\Models\Categorie;
 use App\Models\Conversation;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 
 class CategorieController extends Controller
 {
@@ -17,6 +18,9 @@ class CategorieController extends Controller
     }
     public function create()
     {
+        if(! Gate::allows('create-categorie')){
+            abort(403);
+        }
         $conversations = Conversation::all();
         return view("/back/categories/create", compact("conversations"));
     }
@@ -24,6 +28,7 @@ class CategorieController extends Controller
     {
         
         $categorie = new Categorie;
+        $this->authorize('create', \App\Model\Categorie::class);
         $request->validate([
          'nom'=> 'required',
         ]); // store_validated_anchor;
@@ -33,6 +38,9 @@ class CategorieController extends Controller
     }
     public function edit($id)
     {
+        if(! Gate::allows('update-categorie')){
+            abort(403);
+        }
         $conversations = Conversation::all();
         $categorie = Categorie::find($id);
         return view("/back/categories/edit",compact("categorie" , "conversations"));
@@ -41,6 +49,7 @@ class CategorieController extends Controller
     {
         
         $categorie = Categorie::find($id);
+        $this->authorize('update', \App\Model\Categorie::class);
         $request->validate([
          'nom'=> 'required',
         ]); // update_validated_anchor;
@@ -51,6 +60,7 @@ class CategorieController extends Controller
     public function destroy($id)
     {
         $categorie = Categorie::find($id);
+        $this->authorize('delete', \App\Model\Categorie::class);
         $categorie->delete();
         return redirect()->back()->with('message', "Successful delete !");
     }
