+++ app/Http/Controllers/TagController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers;
 use App\Models\Conversation;
 use App\Models\Tag;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 
 class TagController extends Controller
 {
@@ -17,12 +18,17 @@ class TagController extends Controller
     }
     public function create()
     {
+        if(! Gate::allows('create-tag')){
+            abort(403);
+        }
         $conversations = Conversation::all();
+
         return view("/back/tags/create" , compact("conversations"));
     }
     public function store(Request $request)
     {
         $tag = new Tag;
+        $this->authorize('create', \App\Model\Tag::class);
         $request->validate([
          'nom'=> 'required',
         ]); // store_validated_anchor;
@@ -32,6 +38,9 @@ class TagController extends Controller
     }
     public function edit($id)
     {
+        if(! Gate::allows('update-tag')){
+            abort(403);
+        }
         $conversations = Conversation::all();
         $tag = Tag::find($id);
         return view("/back/tags/edit",compact("tag" , "conversations"));
@@ -39,6 +48,7 @@ class TagController extends Controller
     public function update($id, Request $request)
     {
         $tag = Tag::find($id);
+        $this->authorize('update', \App\Model\Tag::class);
         $request->validate([
          'nom'=> 'required',
         ]); // update_validated_anchor;
@@ -49,6 +59,7 @@ class TagController extends Controller
     public function destroy($id)
     {
         $tag = Tag::find($id);
+        $this->authorize('delete', \App\Model\Tag::class);
         $tag->delete();
         return redirect()->back()->with('message', "Successful delete !");
     }
